feat(ui): Add color correction window with real-time preview
This commit introduces a major new feature: a Color Correction window that
allows users to adjust image properties with a real-time preview.

Key features:
- **UI (`color_correction_window.slint`):**
  - A new window with sliders for Brightness, Contrast, Gamma, RGB Color
    Balance, and Saturation.
  - All adjustment properties are now `float` type, enabling proper
    two-way binding with the `Slider` widgets. This resolves an issue
    where the `Reset` button did not update slider positions.
  - The layout includes "Original" and "Preview" image panes.
- **Real-time Preview (`main.rs`):**
  - On window open, a downscaled thumbnail is generated (preserving
    aspect ratio) and displayed in both preview panes.
  - The `on_values_changed` callback applies adjustments to the thumbnail
    in real-time as sliders are moved, updating the "Preview" pane.
- **Backend Logic (`main.rs`):**
  - The `on_apply` callback applies the final corrections to the
    full-resolution image.
  - `on_cancel` and `on_reset` are fully implemented.

This provides a solid foundation for the color correction functionality.

+++ src/main.rs
@@ -587,8 +587,8 @@ fn main() -> Result<(), slint::PlatformError> {
                 
                 // Apply corrections
                 let brightness = cc_ui.get_brightness();
-                if brightness != 0 {
-                    colorops::brighten_in_place(&mut preview_buffer, brightness);
+                if brightness != 0.0 {
+                    colorops::brighten_in_place(&mut preview_buffer, brightness as i32);
                 }
 
                 let contrast = cc_ui.get_contrast() as f32;
@@ -645,13 +645,13 @@ fn main() -> Result<(), slint::PlatformError> {
     let cc_window_handle = color_correction_window.as_weak();
     color_correction_window.on_reset(move || {
         if let Some(cc_ui) = cc_window_handle.upgrade() {
-            cc_ui.set_brightness(0);
-            cc_ui.set_contrast(0);
-            cc_ui.set_gamma(100);
-            cc_ui.set_red(0);
-            cc_ui.set_green(0);
-            cc_ui.set_blue(0);
-            cc_ui.set_saturation(0);
+            cc_ui.set_brightness(0.0);
+            cc_ui.set_contrast(0.0);
+            cc_ui.set_gamma(100.0);
+            cc_ui.set_red(0.0);
+            cc_ui.set_green(0.0);
+            cc_ui.set_blue(0.0);
+            cc_ui.set_saturation(0.0);
             cc_ui.invoke_values_changed();
         }
     });
@@ -671,7 +671,7 @@ fn main() -> Result<(), slint::PlatformError> {
 
                 // Apply corrections
                 let brightness = cc_ui.get_brightness();
-                if brightness != 0 { colorops::brighten_in_place(&mut buffer, brightness); }
+                if brightness != 0.0 { colorops::brighten_in_place(&mut buffer, brightness as i32); }
 
                 let contrast = cc_ui.get_contrast() as f32;
                 if contrast != 0.0 {
